Bulk add functionality successfully implemented

+++ HelpDeskTool/AddRemoveWindow.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Microsoft.VisualBasic.ApplicationServices;
+using System;
 using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.Linq;
@@ -33,10 +34,12 @@ namespace DTTool
             TextPointer endPointer = rtb.Document.ContentEnd.GetNextInsertionPosition(LogicalDirection.Backward);
             if (startPointer.CompareTo(endPointer) == 0)
             {
-                MessageBox.Show("No text in " + rtb.Name, "Error", MessageBoxButton.OK, MessageBoxImage.Error);
                 return false;
             }
-            return false;
+            else
+            {
+                return true;
+            }
         }
         public string StringFromRichTextBox(RichTextBox rtb)
         {
@@ -52,12 +55,10 @@ namespace DTTool
             return textRange.Text;
         }
 
-        // Remove Button
-        private void RemoveUGButton_Click(object sender, RoutedEventArgs e)
+        public void AddOrRemove(string AddorRemove)
         {
             if (RTPHasText(ARusernameBox) && RTPHasText(ARgroupBox))
             {
-
                 string usertext = StringFromRichTextBox(ARusernameBox);
                 string grouptext = StringFromRichTextBox(ARgroupBox);
                 string errString = "";
@@ -69,12 +70,22 @@ namespace DTTool
                 {
                     foreach (var user in myUserList)
                     {
+                        if (user == "" || group == "")
+                            break;
+
                         group.Trim();
                         user.Trim();
+
                         System.Diagnostics.Process command = new System.Diagnostics.Process();
                         command.StartInfo.CreateNoWindow = true;
                         command.StartInfo.FileName = "cmd";
-                        command.StartInfo.Arguments = "/C powershell Remove-ADGroupMember \'" + group + "\' \'" + user + "\' -Confirm:$false";
+                        if (AddorRemove == "Remove")
+                        {
+                            command.StartInfo.Arguments = "/C powershell Remove-ADGroupMember \'" + group + "\' \'" + user + "\' -Confirm:$false";
+                        }else if(AddorRemove == "Add")
+                        {
+                            command.StartInfo.Arguments = "/C powershell Add-ADGroupMember \'" + group + "\' \'" + user + "\'";
+                        }
                         command.StartInfo.RedirectStandardOutput = true;
                         command.Start();
                         var console_output = command.StandardOutput.ReadToEnd();
@@ -86,50 +97,31 @@ namespace DTTool
                 }
                 if (errString == "")
                 {
-                    MessageBox.Show("All users deleted", "Success", MessageBoxButton.OK, MessageBoxImage.Information);
+                    MessageBox.Show("All users processed successfully\nMay take up to 30 seconds to reflect in AD", "Success", MessageBoxButton.OK, MessageBoxImage.Information);
+                    ARusernameBox.Document.Blocks.Clear();
+                    ARgroupBox.Document.Blocks.Clear();
                 }
                 else
                 {
                     MessageBox.Show(errString, "Error", MessageBoxButton.OK, MessageBoxImage.Error);
                 }
             }
+            else
+            {
+                MessageBox.Show("Input Missing", "Error", MessageBoxButton.OK, MessageBoxImage.Error);
+            }
+        }
+
+        // Remove Button
+        private void RemoveUGButton_Click(object sender, RoutedEventArgs e)
+        {
+            AddOrRemove("Remove");
         }
 
         // Add Button
         private void AddUGButton_Click(object sender, RoutedEventArgs e)
         {
-
-            /* Prevously used code
-                ARoutputbox.Document.Blocks.Clear();
-                var user = ARusernameBox.Text;
-                var group = ARgroupBox.Text;
-                if (user != "" && group != "")
-                {
-                    ARusernameBox.Clear();
-                    ARgroupBox.Clear();
-                    System.Diagnostics.Process command = new System.Diagnostics.Process();
-                    command.StartInfo.CreateNoWindow = true;
-                    command.StartInfo.FileName = "cmd";
-                    // Add-ADGroupMember
-                    // Add-ADGroupMember -Identity GROUP -Members USERNAME
-                    command.StartInfo.Arguments = "/C powershell Add-ADGroupMember \'" + group + "\' \'" + user + "\'";
-                    command.StartInfo.RedirectStandardOutput = true;
-                    command.Start();
-                    var console_output = command.StandardOutput.ReadToEnd();
-                    ARoutputbox.AppendText(console_output);
-                    if (command.ExitCode == 0)
-                    {
-                        ARoutputbox.AppendText(user + " Added to " + group + "\n");
-                        ARoutputbox.AppendText("May take up to 30 seconds for changes to reflect in AD");
-                    }
-                    else
-                    {
-                        ARoutputbox.AppendText("\nAction Failed\n");
-                    }
-                }
-                else ARoutputbox.AppendText("No Input Detected\n");
-                ARoutputbox.ScrollToEnd();
-            */
+            AddOrRemove("Add");
         }
     }
 }
